feat: delete category
