+++ src/validators/storage/products/update.ts
@@ -15,14 +15,6 @@ const schema = yup.object({
     .required('Produto inválido'),
 });
 
-const deleteSchema = yup.object({
-  pid: yup
-    .number()
-    .typeError('Produto inválido')
-    .integer('Produto inválido')
-    .required('Produto inválido'),
-});
-
 export const updateProductValidator = async (req: NextApiRequest) => {
   const { door, floor, categoryId, lockerId } = req.body;
 
@@ -62,22 +54,4 @@ export const updateProductValidator = async (req: NextApiRequest) => {
   if (locker.numberOfFloors < floor) {
     throw new Error('O andar informado não existe');
   }
-};
-
-export const deleteProductValidator = async (req: NextApiRequest) => {
-  if (!(await deleteSchema.isValid(req.query))) {
-    await deleteSchema
-      .validate(req.query, {
-        abortEarly: true,
-      })
-      .catch(err => {
-        throw new Error(err);
-      });
-  }
-
-  const product = await prisma.product.findUnique({
-    where: { id: Number(req.query.pid) },
-  });
-
-  if (!product) throw new Error('Este produto não existe');
 };
